Added NetProbe Beacon and extended properties
Added the Beacon object which will listen to probes that clients will
send when trying to locate running instances of the Qu server on a local
network. This beacon will respond with the information the client
requires in order to subsequently connect to the ClientManager and
become an active connected client. I have also extended the Properties
class to now fetch the Beacon port number and the device access
password, i have also added a number of methods that allows the server
to add/remove super users, as well as get a list of all super users.