Improved ReachableQuDevice and Device
When calling fetchAction on a Device object, if the action is a settings
update then it is processed and updates the Device object before being
passed to the user. Also added super user ids to ReachableQuDevice to
allow the user to determine whether they are a super user just from
getting a probe response.