Started to add JUnit tests, altered NetProbe
Started to add JUnit tests for QuClient API. Also, i've changed the way
we initialise NetProbe objects as i've come to the conclusion that it is
impractical to reference the Android class library in the initialise()
method to get the IP address of the device on the LAN, as we might not
be using the QuClient API from within an Android environment. Instead,
we can pass the address (which the user will have to get themselves as
it can vary so much based on environment) to the initialise method and
the NetProbe object will use that address as the basis for calculating
which address to probe for a server response. We can however pass null
to the initialise method, on this happening NetProbe will attempt to get
the current IP via the java.net.InetAddress class. Got rid of old
QuClient Test class.