Changed the way we initialise Device objects
Changed the way we initialise Device objects. All of the initialisation
and server connection is now done in the link() method and not the
constructor. This is so that now QuEventListeners can subscribe to the
Device BEFORE we connect to the server, removing the risk of vital
events being missed between the initialisation of a Device and the
subscription of a listener. Also, checks have been put in place so that
if the user attempts to invoke any method of a Device object without
calling link() first they will get a RuntimeException. Additionally,
i've added a unit test to check that the QuClient can update settings on
the server, and that the client gets a PUSH_SETTINGS IncomingAction from
the server that subsequently raises a QuEvent.