Worked on QuClient Track file.
Worked on QuClient Track file to implement stop() method. Also added
some basic permissions checking. Fixed issue where play/pause/stop would
throw exception on QuServer if playlist was empty. Added large unit test
in which we test that we can upload a track, that it plays, that we can
pause it, that we can play it again, and that we can stop it.