Increased timeout for welcome package
Increased timeout for welcome package on client as for some reason when
we run the server on a raspberry pi 2 and the client on a linux box the
wait between client connection and the getting of the welcome package is
always suspiciously close to 2000ms. there might be an issue here but
for now this will do.