Major cleanup for QuClient
Major cleanup for QuClient. Also, it turns out that Android has issues
with doing any networking on its main thread. Because of this i've had
to put various parts of the ActionChannel/Listener in a different
thread. As the Android app is being developed there will most likely be
more changes to be made to the client. All QuClient unit tests pass.